#OSIS-1783 Modifications review
