+++ base/locale/en/LC_MESSAGES/django.po
@@ -3878,7 +3878,7 @@ msgstr ""
 msgid "education group detailed achievement"
 msgstr ""
 
-msgid "Produce configurable xls of learning units"
+msgid "Produce a customizable xls of learning units"
 msgstr ""
 
 msgid "Configurable list of learning units"
@@ -3887,7 +3887,7 @@ msgstr ""
 msgid "Do you want to export the following informations"
 msgstr ""
 
-msgid "With programs"
+msgid "Programs/gathering"
 msgstr ""
 
 msgid "List of teachers"
@@ -3918,7 +3918,7 @@ msgid "2nd quadri"
 msgstr ""
 
 msgid "Suppression proposal"
-msgstr "Suppression proposal"
+msgstr ""
 
 msgid "Transformation proposal"
 msgstr ""
