+++ base/templates/learning_unit/blocks/modal/modal_xls_with_parameters.html
@@ -35,7 +35,7 @@
         <div class="modal-body">
             {% csrf_token %}
             <label>{% trans 'Do you want to export the following informations' %}?</label><br>
-            <input type="checkbox" id="chb_with_grp" value="chb_with_grp"> {% trans 'With programs' %}<br>
+            <input type="checkbox" id="chb_with_grp" value="chb_with_grp"> {% trans 'Programs/gathering' %}<br>
             <input type="checkbox" id="chb_with_attributions" value="chb_with_attributions"> {% trans 'tutors' %}<br>
         </div>
         <div class="modal-footer">
