+++ base/templates/learning_unit/by_activity.html
@@ -63,7 +63,7 @@
                     <li>
                         <a id="btn_produce_xls_with_parameters"
                            style="margin-right:10px;"
-                           title="{% trans 'Produce configurable xls of learning units' %}"
+                           title="{% trans 'Produce a customizable xls of learning units' %}"
                            href="#"
                            data-toggle="modal" data-target="#modalXlsLuyWithParameters">
                             &nbsp;{% trans 'Configurable list of learning units' %}
